chore: refactor service responses to extend `ApplicationCall`
Made in preparation for Server Sent Events, which do not implement the `RoutingContext` and therefore wouldn't be able to use the existing functions.
