+++ src/main/kotlin/me/snoty/backend/server/resources/wiring/flow/_FlowResourcesUtils.kt
@@ -1,5 +1,6 @@
 package me.snoty.backend.server.resources.wiring.flow
 
+import io.ktor.server.application.*
 import io.ktor.server.routing.*
 import me.snoty.backend.server.plugins.void
 import me.snoty.backend.utils.getUser
@@ -9,9 +10,11 @@ import me.snoty.integration.common.wiring.flow.FlowService
 import me.snoty.integration.common.wiring.flow.StandaloneWorkflow
 import org.koin.ktor.ext.get
 
-suspend fun RoutingContext.getPersonalFlowOrNull(): StandaloneWorkflow? {
-	val user = call.getUser()
-	val id = call.parameters["id"] ?: return void { invalidNodeId() }
+suspend fun RoutingContext.getPersonalFlowOrNull() = call.getPersonalFlowOrNull()
+
+suspend fun ApplicationCall.getPersonalFlowOrNull(): StandaloneWorkflow? {
+	val user = getUser()
+	val id = parameters["id"] ?: return void { invalidNodeId() }
 
 	val flow = get<FlowService>().getStandalone(id)
 	if (flow?.userId != user.id) {
